requirements.txt
Pillow==7.2.0
playsound==1.2.2
PySimpleGUI==4.28.0

